+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v0.16
 
+### v0.16.4
+
+- Fixes disconnect not working. #122.
+- Auto connecting if have just one connection
+- 'Add new Server' changed to 'Add new connection'
+
 ### v0.16.3
 
 - Show records and describe tables when using multiple connections. #119